All homeWork

+++ Task_021_HomeWork/Program.cs
@@ -7,14 +7,14 @@ A (3,6,8); B (2,1,-7), -> 15.84
 A (7,-5, 0); B (1,-1,9) -> 11.53*/
 
 System.Console.WriteLine("Введите 1 координ");
-double x1 = Convert.ToInt32(Console.ReadLine());
-double y1 = Convert.ToInt32(Console.ReadLine());
-double z1 = Convert.ToInt32(Console.ReadLine());
+double x1 = Convert.ToDouble(Console.ReadLine()); 
+double y1 = Convert.ToDouble(Console.ReadLine());
+double z1 = Convert.ToDouble(Console.ReadLine());
 
 System.Console.WriteLine("Введите 2 координ");
-double x2 = Convert.ToInt32(Console.ReadLine());
-double y2 = Convert.ToInt32(Console.ReadLine());
-double z2 = Convert.ToInt32(Console.ReadLine());
+double x2 = Convert.ToDouble(Console.ReadLine());
+double y2 = Convert.ToDouble(Console.ReadLine());
+double z2 = Convert.ToDouble(Console.ReadLine());
 
 double res = Hypotenuse(x1,y1,x2,y2,z1,z2);
 Console.WriteLine("Result => " + Math.Round(res,2,MidpointRounding.ToZero));
